Common testing code extracted to separate file
